BUGFIX: Fix non-contiguous Exit-assignment towards channel exit.
* We now use ComponentPositionComparator in ComponentTreeUtils.isAbove().
* We do this, because the previous version is broken, now that we have
2D components, where the bottom-edge of the top-component can be below
the top-edge of the bottom component. Therefore the condition
candMinMax.getB() < refMinMax.getA();
is no longer sufficient to test, if one component is above the other.
* This fixes the bug, where we had non-contiguous Exit-assignment towards
the exit-of the channel because the constraint of Eq.8, in of the paper
"Optimal Joint Segmentation and Tracking of Escherichia Coli in the Mother Machine"
was not being set correctly anymore.